+++ src/utils/validator.ts
@@ -75,7 +75,7 @@ export function isNumber(value: any): boolean {
  * @param {any} value The value to validate.
  * @return {boolean} Whether the value is a string or not.
  */
-export function isString(value: any): boolean {
+export function isString(value: any): value is string {
   return typeof value === 'string';
 }
 
@@ -100,7 +100,7 @@ export function isBase64String(value: any): boolean {
  * @param {any} value The value to validate.
  * @return {boolean} Whether the value is a non-empty string or not.
  */
-export function isNonEmptyString(value: any): boolean {
+export function isNonEmptyString(value: any): value is string {
   return isString(value) && value !== '';
 }