+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Unreleased
 
+- [fixed] Correctly parses error codes with details messages in Firebase Auth.
+- [fixed] Fixed optional fields in UserRecord types to be optional.
+- [added] `admin.projectManagement().shaCertificate()` method to create an
+  instance of admin.projectManagement.ShaCertificate.
+
+# v6.4.0
+
+- [added] `messaging.Aps` type now supports configuring a critical sound.
+  A new `messaging.CriticalSound` type has been introduced for this purpose.
+- [added] `messaging.AndroidNotification` type now supports `channel_id`.
 - [added] `AppOptions` now accepts an optional `http.Agent` object. The
   `http.Agent` specified via this API is used when the SDK makes backend
   HTTP calls. This can be used when it is required to deploy the Admin SDK
@@ -8,9 +18,6 @@
   and `admin.credential.refreshToken()` methods now accept an `http.Agent`
   as an optional argument. If specified, the `http.Agent` will be used
   when calling Google backend servers to fetch OAuth2 access tokens.
-- [added] `messaging.AndroidNotification` type now supports channel_id.
-- [added] `admin.projectManagement().shaCertificate()` method to create an
-  instance of admin.projectManagement.ShaCertificate.
 
 # v6.3.0